+++ src/clojure2d/color.clj
@@ -5415,7 +5415,6 @@
 
 (m/unuse-primitive-operators)
 
-
 (comment (run! (fn [[k [to from]]]
                  (println "---")
                  (println k)